[MINOR] Frame Float detection refinement
This commit refine the detection and selection of float values in the
schema detection algorithm located in:

src/main/java/org/apache/sysds/runtime/frame/data/lib/FrameUtil.java

To improve the performance a custom matcher have been made to avoid using
regexes if possible, and we try to avoid parsing the float value from the
string representation if at all possible. The implementation is not
completely fool proof and does not consider many adversarial inputs.

on a small test case of 1000 string values that all are fp32 the
implementation improve performance from 0.5 ms to 0.02 ms.

+++ src/test/java/org/apache/sysds/test/component/frame/FrameUtilTest.java
@@ -129,17 +129,17 @@ public class FrameUtilTest {
 
 	@Test
 	public void testInfinite() {
-		assertEquals(ValueType.FP64, FrameUtil.isType("infinity"));
+		assertEquals(ValueType.FP32, FrameUtil.isType("infinity"));
 	}
 
 	@Test
 	public void testMinusInfinite() {
-		assertEquals(ValueType.FP64, FrameUtil.isType("-infinity"));
+		assertEquals(ValueType.FP32, FrameUtil.isType("-infinity"));
 	}
 
 	@Test
 	public void testNan() {
-		assertEquals(ValueType.FP64, FrameUtil.isType("nan"));
+		assertEquals(ValueType.FP32, FrameUtil.isType("nan"));
 	}
 
 	@Test
